os: Implement `os::Event`
