+++ CMakeLists.txt
@@ -171,6 +171,9 @@ add_library(NintendoSDK OBJECT
   include/nn/oe.h
   include/nn/util.h
   include/nn/image.h
+  include/nn/os/os_Event.h
+  include/nn/os/os_EventCommon.h
+  include/nn/os/os_EventTypes.h
   include/nn/util/AccessorBase.h
   include/nn/util/MathTypes.h
   include/nn/util/util_Arithmetic.h
