+++ include/nn/os.h
@@ -11,9 +11,9 @@
 #include <nn/types.h>
 
 #include <nn/os/detail/os_InternalCriticalSection.h>
+#include <nn/os/os_Event.h>
 #include <nn/os/os_MessageQueueTypes.h>
 #include <nn/os/os_Mutex.h>
-#include <nn/os/os_MutexTypes.h>
 #include <nn/os/os_ThreadTypes.h>
 
 namespace nn {
@@ -21,7 +21,6 @@ namespace os {
 
 namespace detail {
 
-class MultiWaitObjectList;
 struct InterProcessEventType {
     enum State {
         State_NotInitialized = 0,
@@ -48,22 +47,6 @@ struct Tick {
 struct LightEventType {
     std::aligned_storage_t<0xc, 4> storage;
 };
-
-struct EventType {
-    util::TypedStorage<detail::MultiWaitObjectList, 16, 8> _multiWaitObjectList;
-    bool _signalState;
-    bool _initiallySignaled;
-    uint8_t _clearMode;
-    uint8_t _state;
-    uint32_t _broadcastCounterLower;
-    uint32_t _broadcastCounterUpper;
-    detail::InternalCriticalSectionStorage _csEvent;
-    detail::InternalConditionVariableStorage _cvSignaled;
-};
-static_assert(std::is_trivial<EventType>::value, "EventType non trivial");
-typedef EventType Event;
-
-enum EventClearMode { EventClearMode_ManualClear, EventClearMode_AutoClear };
 
 struct ConditionVariableType {};
 
@@ -153,15 +136,6 @@ void ResumeThread(nn::os::ThreadType*);
 void SleepThread(nn::TimeSpan);
 void WaitThread(nn::os::ThreadType*);
 void SetThreadCoreMask(nn::os::ThreadType*, int, u64 mask);
-
-// EVENTS
-void InitializeEvent(EventType*, bool initiallySignaled, EventClearMode eventClearMode);
-void FinalizeEvent(EventType*);
-void SignalEvent(EventType*);
-void WaitEvent(EventType*);
-bool TryWaitEvent(EventType*);
-bool TimedWaitEvent(EventType*, nn::TimeSpan);
-void ClearEvent(EventType*);
 
 // LIGHT EVENTS
 void InitializeLightEvent(LightEventType*, bool initiallySignaled, EventClearMode eventClearMode);
