+++ include/nn/os/os_EventCommon.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace nn::os {
+enum EventClearMode { EventClearMode_ManualClear, EventClearMode_AutoClear };
+}  // namespace nn::os
